Fix size parameter in header file. JIRA: DEV-5179

+++ channels/smartcard/client/smartcard_main.h
@@ -127,7 +127,7 @@ UINT smartcard_process_irp(SMARTCARD_DEVICE* smartcard, IRP* irp);
 
 FREERDP_API DEVICE * awingu_smartcard_entry(DEVMAN * devman);
 FREERDP_API int awingu_process_irp(IRP * irp);
-FREERDP_API int awingu_onmessage(BYTE * buffer, size_t size, DEVMAN * devman);
+FREERDP_API int awingu_onmessage(BYTE * buffer, int size, DEVMAN * devman);
 
 LONG smartcard_irp_device_control_decode(SMARTCARD_DEVICE* smartcard, SMARTCARD_OPERATION* operation);
 LONG smartcard_irp_device_control_call(SMARTCARD_DEVICE* smartcard, SMARTCARD_OPERATION* operation);
